feat: deprecate envvar segment
favors text in a more generic way
