+++ docs/docs/segment-text.md
@@ -22,11 +22,23 @@ Display text.
 }
 ```
 
+:::tip
+If you're using PowerShell, you can override a function to populate environment variables before the
+prompt is rendered.
+
+```powershell
+function Set-EnvVar {
+    $env:POSH=$(Get-Date)
+}
+New-Alias -Name 'Set-PoshContext' -Value 'Set-EnvVar' -Scope Global -Force
+```
+
+:::
+
 ## Properties
 
-- text: `string` - text/icon to display. Accepts [coloring foreground][coloring] just like `prefix` and `postfix`.
-Powered by go [text/template][go-text-template] templates extended with [sprig][sprig] utilizing the
-properties below.
+- template: `string` - text/icon to display. Powered by go [text/template][go-text-template] templates extended
+with [sprig][sprig] utilizing the properties below.
 
 ## Template Properties
