+++ src/segment_deprecated_test.go
@@ -762,3 +762,34 @@ func TestPythonVirtualEnv(t *testing.T) {
 		assert.Equal(t, tc.Expected, python.string(), tc.Case)
 	}
 }
+
+// Environment Variable
+
+func TestEnvvarAvailable(t *testing.T) {
+	name := "HERP"
+	expected := "derp"
+	env := new(MockedEnvironment)
+	env.On("getenv", name).Return(expected)
+	e := &envvar{
+		env: env,
+		props: map[Property]interface{}{
+			VarName: name,
+		},
+	}
+	assert.True(t, e.enabled())
+	assert.Equal(t, expected, e.string())
+}
+
+func TestEnvvarNotAvailable(t *testing.T) {
+	name := "HERP"
+	expected := ""
+	env := new(MockedEnvironment)
+	env.On("getenv", name).Return(expected)
+	e := &envvar{
+		env: env,
+		props: map[Property]interface{}{
+			VarName: name,
+		},
+	}
+	assert.False(t, e.enabled())
+}
